requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
cloudpickle==0.6.1
colorama==0.4.1
configparser==3.5.0
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
enum34==1.1.6
futures==3.2.0
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==2.2.3
mccabe==0.6.1
networkx==2.2
numpy==1.15.4
Pillow==5.4.0
pylint==1.9.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
scikit-fuzzy==0.4.0
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
toolz==0.9.0
wrapt==1.10.11
